# apps/remote/openapi.yaml
openapi: 3.0.3
info:
  title: Remote data fields API
  version: 0.1.0
  description: >-
    A stateless, read-only decision service for one ordered form. Clients hold
    draft state and use this API to discover the next field, search strict
    selectable values, and validate the final state before separately saving it.
paths:
  /form/step:
    post:
      operationId: evaluateFormStep
      summary: Evaluate the next form step
      description: >-
        Validates `state` as a complete, contiguous answer prefix for the
        current form path. Returns the next field to confirm or a completion
        result. The server owns all field keys, field order, skipped fields,
        and completion rules.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormStateRequest"
      responses:
        "200":
          description: The next applicable field or a completed form path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormStep"
        "422":
          $ref: "#/components/responses/InvalidFormState"
  /fields/{fieldKey}/options:
    post:
      operationId: searchFieldOptions
      summary: Search values for a field
      description: >-
        Returns up to 25 server-ranked values for the current field in the
        supplied confirmed-state context. The field must be the next field
        reachable from `state`; returned values are strict selectable answers.
      parameters:
        - $ref: "#/components/parameters/FieldKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionSearchRequest"
      responses:
        "200":
          description: Matching values, limited to 25 options.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptionSearchResult"
        "404":
          $ref: "#/components/responses/UnknownField"
        "422":
          $ref: "#/components/responses/InvalidFormState"
  /form/validate:
    post:
      operationId: validateFormState
      summary: Validate a completed form
      description: >-
        Validates the complete form state against current rules and source
        data immediately before the client persists it. The API never saves
        form state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormStateRequest"
      responses:
        "204":
          description: The completed form state is currently valid.
        "422":
          $ref: "#/components/responses/InvalidFormState"
components:
  parameters:
    FieldKey:
      name: fieldKey
      in: path
      required: true
      description: >-
        Field identifier returned by the current `POST /form/step` response.
        The identifier is valid only for the supplied confirmed-state prefix.
      schema:
        type: string
        minLength: 1
  responses:
    InvalidFormState:
      description: >-
        The state is structurally valid JSON but is not a valid reachable form
        state under the current rules or source data.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/FormStateProblem"
    UnknownField:
      description: The requested field key is not defined by this form.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
    FormState:
      type: object
      description: >-
        Confirmed, contiguous answer prefix for the current form path. A
        missing key is unanswered or not applicable. `null` is an explicit
        no-value answer and is accepted only for a field that permits it.
      additionalProperties:
        type: string
        nullable: true
    FormStateRequest:
      type: object
      required: [state]
      properties:
        state:
          $ref: "#/components/schemas/FormState"
      additionalProperties: false
    OptionSearchRequest:
      type: object
      required: [state, query]
      properties:
        state:
          $ref: "#/components/schemas/FormState"
        query:
          type: string
          description: >-
            Text typed into the current field. The field descriptor specifies
            the minimum permitted length.
      additionalProperties: false
    OptionSearchResult:
      type: object
      required: [options, truncated]
      properties:
        options:
          type: array
          description: >-
            Strict selectable values, ordered by the server's field-specific
            relevance rules. Each value is both the persisted value and its
            display label.
          maxItems: 25
          items:
            type: string
        truncated:
          type: boolean
          description: >-
            `true` when additional matches exist beyond the returned 25.
            Clients should ask the user to refine the query instead of paging.
      additionalProperties: false
    FormStep:
      oneOf:
        - $ref: "#/components/schemas/NextFieldStep"
        - $ref: "#/components/schemas/CompleteFormStep"
    NextFieldStep:
      type: object
      required: [state, complete, field]
      properties:
        state:
          $ref: "#/components/schemas/FormState"
        complete:
          type: boolean
          enum: [false]
        field:
          $ref: "#/components/schemas/Field"
      additionalProperties: false
    CompleteFormStep:
      type: object
      required: [state, complete]
      properties:
        state:
          $ref: "#/components/schemas/FormState"
        complete:
          type: boolean
          enum: [true]
      additionalProperties: false
    Field:
      description: Current field descriptor returned by the form decision service.
      oneOf:
        - $ref: "#/components/schemas/RequiredValueField"
        - $ref: "#/components/schemas/NullableValueField"
    RequiredValueField:
      type: object
      required: [key, label, minimumQueryLength, allowsNull]
      properties:
        key:
          type: string
          minLength: 1
          description: Identifier to use in form state and option-search paths.
        label:
          type: string
          minLength: 1
          description: User-visible label for the current field.
        description:
          type: string
          description: Optional user-visible help text for the current field.
        placeholder:
          type: string
          description: Optional prompt shown in the type-ahead input.
        minimumQueryLength:
          type: integer
          minimum: 0
          description: Minimum typed characters before the client searches.
        allowsNull:
          type: boolean
          enum: [false]
          description: An explicit `null` answer cannot advance this step.
      additionalProperties: false
    NullableValueField:
      type: object
      required: [key, label, minimumQueryLength, allowsNull, nullLabel]
      properties:
        key:
          type: string
          minLength: 1
          description: Identifier to use in form state and option-search paths.
        label:
          type: string
          minLength: 1
          description: User-visible label for the current field.
        description:
          type: string
          description: Optional user-visible help text for the current field.
        placeholder:
          type: string
          description: Optional prompt shown in the type-ahead input.
        minimumQueryLength:
          type: integer
          minimum: 0
          description: Minimum typed characters before the client searches.
        allowsNull:
          type: boolean
          enum: [true]
          description: An explicit `null` answer may advance this step.
        nullLabel:
          type: string
          minLength: 1
          description: User-visible label for the explicit `null` value.
      additionalProperties: false
    ProblemDetails:
      type: object
      description: RFC 9457 problem details response.
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable machine-readable problem identifier.
        title:
          type: string
          description: Short summary of the problem type.
        status:
          type: integer
          description: HTTP status code generated by the origin server.
        detail:
          type: string
          description: User-visible explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this problem occurrence.
        timestamp:
          type: string
          format: date-time
          description: Time when the server generated the problem.
      additionalProperties: true
    FormStateProblem:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          required: [type, title, status, detail, errors]
          properties:
            firstInvalidFieldKey:
              type: string
              description: >-
                Earliest field requiring correction. Omitted when the problem
                cannot be associated with a submitted field.
            errors:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/FormStateError"
    FormStateError:
      type: object
      required: [type, detail]
      properties:
        fieldKey:
          type: string
          description: Field associated with this problem when applicable.
        type:
          type: string
          format: uri-reference
          description: Stable machine-readable problem identifier.
        detail:
          type: string
          description: User-visible explanation specific to this field.
      additionalProperties: false